backend/ai_engine/gemini_client.py
import os
from google import genai

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

def generate_text(prompt: str) -> str:
    response = client.models.generate_content(
        model="gemini-1.0-pro",
        contents=prompt,
    )
    return response.text.strip()
